requirements.txt
asgiref==3.2.10
Django==3.0.8
django-crispy-forms==1.9.1
django-widget-tweaks==1.4.8
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
djangorestframework==3.11.0
Markdown==3.2.2
filter==0.0.0.20200724
channels==2.0.4